requirements.txt
pypdf
streamlit
google-generativeai
numpy
